move part to azure_part
