+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/configs/urls.py
@@ -20,18 +20,17 @@ from django.contrib import admin
 from django.urls import path
 from rest_framework import routers
 
-from vision_on_edge.azure_settings.api import views as setting_views
+from vision_on_edge.azure_parts.api import views as azure_part_views
+from vision_on_edge.azure_settings.api import views as azure_setting_views
 from vision_on_edge.azure_training.api import views as azure_training_views
 from vision_on_edge.cameras.api import util_views as camera_util_views
 from vision_on_edge.cameras.api import views
 from vision_on_edge.image_predictions.api import \
     views as image_prediction_views
-from vision_on_edge.locations.api import views as location_views
-from vision_on_edge.part.api import views as part_views
-from vision_on_edge.streams.api import views as stream_views
 from vision_on_edge.images.api import views as image_views
+from vision_on_edge.locations.api import views as location_views
 from vision_on_edge.relabeling.api import views as relabel_views
-
+from vision_on_edge.streams.api import views as stream_views
 
 from . import views as site_views
 
@@ -46,9 +45,9 @@ class OptionalSlashRouter(routers.DefaultRouter):
 
 #router = ters.DefaultRouter(trailing_slash=False)
 router = OptionalSlashRouter()
-router.register('settings', setting_views.SettingViewSet)
+router.register('settings', azure_setting_views.SettingViewSet)
 router.register('cameras', views.CameraViewSet)
-router.register('parts', part_views.PartViewSet)
+router.register('parts', azure_part_views.PartViewSet)
 router.register('locations', location_views.LocationViewSet)
 router.register('image_predictions',
                 image_prediction_views.ImagePredictionViewSet)
