+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_parts/__init__.py
@@ -1,0 +1,5 @@
+"""
+vision_on_edge.azure_parts.
+Application for azure training part.
+"""
+default_app_config = 'vision_on_edge.azure_parts.apps.AzurePartsConfig'
