+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_parts/api/views.py
@@ -1,5 +1,5 @@
 """
-Part views
+Azure Parts views
 """
 from __future__ import absolute_import, unicode_literals
