+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_parts/models.py
