+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_parts/tests/test_model_part.py
