+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_training/models.py
@@ -13,11 +13,11 @@ from azure.iot.device import IoTHubModuleClient
 from django.db import models
 from django.db.models.signals import post_save, pre_save
 
+from ..azure_parts.models import Part
 from ..azure_settings.models import Setting
 from ..cameras.models import Camera
 from ..images.models import Image
 from ..locations.models import Location
-from ..part.models import Part
 from .utils.app_insight import (get_app_insight_logger, img_monitor,
                                 part_monitor, retraining_job_monitor,
                                 training_job_monitor)
