+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/images/models.py
@@ -15,7 +15,7 @@ from django.db.models.signals import pre_delete, pre_save
 from PIL import Image as PILImage
 from rest_framework import status
 
-from ..part.models import Part
+from ..azure_parts.models import Part
 
 logger = logging.getLogger(__name__)
